+++ examples/random_how_to.py
@@ -1,0 +1,4 @@
+from pywikihow import RandomHowTo
+
+how_to = RandomHowTo()
+how_to.print()