+++ pywikihow/exceptions.py
@@ -1,0 +1,4 @@
+
+
+class ParseError(RuntimeError):
+    """ Error parsing WikiHow page"""