+++ CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0]  - 2019-12-12
+
+Breaking Changes, api is backward incompatible
+
+### Changed
+
+- Refactor into individual classes
+    - WikiHow
+    - HowTo
+    - HowToStep
+- Refactor search
+    - WikiHow.search is now a generator
+    - search now returns HowTo objects instead of urls
+- split examples in several files
+
+### Added
+
+- search_wikihow function
+- ParseError exception
+
+### Fixed
+
+- handle parse errors
+
 ## [0.3.1]  - 2019-12-12
 
 ### Changed
@@ -15,4 +39,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made a changelog
 
 [unreleased]: https://github.com/OpenJarbas/PyWikiHow/tree/dev
+[0.5.0]: https://github.com/OpenJarbas/PyWikiHow/tree/0.5.0
 [0.3.1]: https://github.com/OpenJarbas/PyWikiHow/tree/0.3.1